requirements.txt
pygame-ce>=2.3.2
pygame_wrapper>=0.2.1